+++ exporters/otlp/src/otlp_http_log_record_exporter.cc
@@ -81,9 +81,16 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
 #ifdef ENABLE_ASYNC_EXPORT
           ,
           options.max_concurrent_requests,
-          options.max_requests_per_connection
+          options.max_requests_per_connection,
+#else
+          ,
+          64,
+          8,
 #endif
-          )))
+          GetOtlpDefaultUserAgent(),
+          options.http_keepalive,
+          options.http_keepalive_idle,
+          options.http_keepalive_intvl)))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -119,9 +126,16 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
 #ifdef ENABLE_ASYNC_EXPORT
           ,
           options.max_concurrent_requests,
-          options.max_requests_per_connection
+          options.max_requests_per_connection,
+#else
+          ,
+          64,
+          8,
 #endif
-          )))
+          GetOtlpDefaultUserAgent(),
+          options.http_keepalive,
+          options.http_keepalive_idle,
+          options.http_keepalive_intvl)))
 {}
 
 OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
@@ -157,9 +171,16 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
 #ifdef ENABLE_ASYNC_EXPORT
                                 ,
                                 options.max_concurrent_requests,
-                                options.max_requests_per_connection
+                                options.max_requests_per_connection,
+#else
+                                ,
+                                64,
+                                8,
 #endif
-                                ),
+                                GetOtlpDefaultUserAgent(),
+                                options.http_keepalive,
+                                options.http_keepalive_idle,
+                                options.http_keepalive_intvl),
           factory))
 {}
 
@@ -197,9 +218,16 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
 #ifdef ENABLE_ASYNC_EXPORT
                                 ,
                                 options.max_concurrent_requests,
-                                options.max_requests_per_connection
+                                options.max_requests_per_connection,
+#else
+                                ,
+                                64,
+                                8,
 #endif
-                                ),
+                                GetOtlpDefaultUserAgent(),
+                                options.http_keepalive,
+                                options.http_keepalive_idle,
+                                options.http_keepalive_intvl),
           factory))
 {}
 
@@ -236,9 +264,16 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
 #ifdef ENABLE_ASYNC_EXPORT
                                 ,
                                 options.max_concurrent_requests,
-                                options.max_requests_per_connection
+                                options.max_requests_per_connection,
+#else
+                                ,
+                                64,
+                                8,
 #endif
-                                ),
+                                GetOtlpDefaultUserAgent(),
+                                options.http_keepalive,
+                                options.http_keepalive_idle,
+                                options.http_keepalive_intvl),
           std::move(http_client)))
 {}
 
@@ -276,9 +311,16 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(
 #ifdef ENABLE_ASYNC_EXPORT
                                 ,
                                 options.max_concurrent_requests,
-                                options.max_requests_per_connection
+                                options.max_requests_per_connection,
+#else
+                                ,
+                                64,
+                                8,
 #endif
-                                ),
+                                GetOtlpDefaultUserAgent(),
+                                options.http_keepalive,
+                                options.http_keepalive_idle,
+                                options.http_keepalive_intvl),
           std::move(http_client)))
 {}
 
@@ -302,6 +344,9 @@ OtlpHttpLogRecordExporter::OtlpHttpLogRecordExporter(std::unique_ptr<OtlpHttpCli
   options_.max_requests_per_connection = http_client_->GetOptions().max_requests_per_connection;
 #endif
   runtime_options_.thread_instrumentation = http_client_->GetOptions().thread_instrumentation;
+  options_.http_keepalive       = http_client_->GetOptions().http_keepalive;
+  options_.http_keepalive_idle  = http_client_->GetOptions().http_keepalive_idle;
+  options_.http_keepalive_intvl = http_client_->GetOptions().http_keepalive_intvl;
 }
 // ----------------------------- Exporter methods ------------------------------